Add files via upload
I've done a simple search feature based only on matching the genres with the titles to demonstrate how it'll look. On the search by title page, when you click on a recommendation, a pop up will show the description from the TMDB site.

And opinions on the look and feel of the site are more than welcome. I obviously haven't though too much about that yet